add Pollard's rho algorithm

+++ ACM-Codebook/Book2 - Number Theory/number-theory/powmod.cpp
@@ -1,4 +1,4 @@
-LL powmod(LL b, LL e, LL m){
+LL powmod(LL b, unsigned long long e, LL m){
     LL r = 1;
     while (e){
         if (e & 1) r = r * b % m;
